Factor out simulate_npt_traj, merge examples/npt_*.py
