--- examples/npt_simulation.py
@@ -1,137 +1,0 @@
-# Run a simulation at constant temperature and pressure
-
-import numpy as np
-from simtk import unit
-from tqdm import tqdm
-
-from testsystems.relative import hif2a_ligand_pair
-
-from md.builders import build_water_system
-from md.minimizer import minimize_host_4d
-
-from fe.topology import SingleTopology
-from fe.free_energy import RelativeFreeEnergy
-
-from md.ensembles import PotentialEnergyModel, NPTEnsemble
-from md.barostat.moves import MonteCarloBarostat, CoordsAndBox
-from md.barostat.utils import get_group_indices, compute_box_volume
-from md.thermostat.utils import sample_velocities, run_thermostatted_md
-
-from timemachine.lib import LangevinIntegrator
-
-if __name__ == '__main__':
-
-    # thermodynamic parameters
-    temperature = 300 * unit.kelvin
-    pressure = 1.0 * unit.atmosphere
-
-    # build a pair of alchemical ligands in a water box
-    mol_a, mol_b, core, ff = hif2a_ligand_pair.mol_a, hif2a_ligand_pair.mol_b, hif2a_ligand_pair.core, hif2a_ligand_pair.ff
-    complex_system, complex_coords, complex_box, complex_top = build_water_system(4.0)
-
-    min_complex_coords = minimize_host_4d([mol_a, mol_b], complex_system, complex_coords, ff, complex_box)
-
-    single_topology = SingleTopology(mol_a, mol_b, core, ff)
-    rfe = RelativeFreeEnergy(single_topology)
-
-    unbound_potentials, sys_params, masses, coords = rfe.prepare_host_edge(
-        ff.get_ordered_params(), complex_system, min_complex_coords
-    )
-
-    # define NPT ensemble
-    potential_energy_model = PotentialEnergyModel(sys_params, unbound_potentials)
-    ensemble = NPTEnsemble(potential_energy_model, temperature, pressure)
-
-    lam = 0.63
-
-
-    def reduced_potential_fxn(x, box):
-        u, du_dx = ensemble.reduced_potential_and_gradient(x, box, lam)
-        return u
-
-
-    # define a barostat
-
-    # (getting list of molecules is most easily done by looking at bond table)
-    harmonic_bond_potential = unbound_potentials[0]
-    group_indices = get_group_indices(harmonic_bond_potential)
-
-    barostat = MonteCarloBarostat(reduced_potential_fxn, group_indices, max_delta_volume=3.0)
-
-    # define a thermostat
-    seed = 2021
-    integrator = LangevinIntegrator(
-        temperature.value_in_unit(unit.kelvin),
-        1.5e-3,
-        1.0,
-        masses,
-        seed
-    )
-    integrator_impl = integrator.impl()
-
-    def simulate_npt_traj(coords, box, n_moves=1000):
-        barostat.reset()
-
-        # alternate between thermostat moves and barostat moves
-        traj = [CoordsAndBox(coords, box)]
-        volume_traj = [compute_box_volume(traj[0].box)]
-        proposal_scale_traj = [barostat.max_delta_volume]
-
-        trange = tqdm(range(n_moves))
-
-        from time import time
-
-        v_t = sample_velocities(masses * unit.amu, temperature)
-
-        for _ in trange:
-            t0 = time()
-            x_0, v_0, box = traj[-1].coords, v_t.copy(), traj[-1].box
-            x_t, v_t = run_thermostatted_md(integrator_impl, potential_energy_model.all_impls, x_0, box, v_0, lam,
-                                            n_steps=5)
-            after_nvt = CoordsAndBox(x_t, box)
-            t1 = time()
-            after_npt = barostat.move(after_nvt)
-            t2 = time()
-
-            traj.append(after_npt)
-            volume_traj.append(compute_box_volume(after_npt.box))
-            proposal_scale_traj.append(barostat.max_delta_volume)
-
-            trange.set_postfix(volume=f'{volume_traj[-1]:.3f}',
-                               acceptance_fraction=f'{barostat.acceptance_fraction:.3f}',
-                               md_proposal_time=f'{(t1 - t0):.3f}s',
-                               barostat_proposal_time=f'{(t2 - t1):.3f}s',
-                               proposal_scale=f'{barostat.max_delta_volume:.3f}',
-                               )
-
-        traj = np.array(traj)
-        volume_traj = np.array(volume_traj)
-        return traj, volume_traj, proposal_scale_traj
-
-
-    import matplotlib.pyplot as plt
-
-    initial_box_scales = np.array([1.0, 1.05, 1.1, 1.15, 1.2, 1.25]) ** (1.0 / 3)
-    trajs = []
-    volume_trajs = []
-    proposal_scale_trajs = []
-    for scale in initial_box_scales[::-1]:
-        traj, volume_traj, proposal_scale_traj = simulate_npt_traj(coords, complex_box * scale, n_moves=2000)
-
-        trajs.append(traj)
-        volume_trajs.append(volume_traj)
-        proposal_scale_trajs.append(proposal_scale_traj)
-
-    for volume_traj in volume_trajs[::-1]:
-        plt.plot(volume_traj)
-    plt.xlabel('# moves')
-    plt.ylabel('volume')
-    plt.savefig('volume_traj.png', dpi=300, bbox_inches='tight')
-    plt.close()
-
-    for proposal_scale_traj in proposal_scale_trajs[::-1]:
-        plt.plot(proposal_scale_traj)
-    plt.xlabel('# moves')
-    plt.ylabel('proposal scale')
-    plt.savefig('proposal_scale_traj.png', dpi=300, bbox_inches='tight')
-    plt.close()
