+++ examples/readme.md
@@ -1,8 +1,7 @@
 ### Examples
 
 #### MD
-`npt_simulation.py` samples the constant temperature and pressure ensemble by running thermostatted MD and pausing every few steps to apply a Monte Carlo barostat move
-`npt_equilibration_at_multiple_lambdas.py` runs a simulation including a barostat on several lambda windows
+`npt_equilibration.py` samples the constant temperature and pressure ensemble for each of several lambda windows, by running thermostatted MD and pausing every few steps to apply a Monte Carlo barostat move
 
 #### Absolute hydration free energy
